\name{example.map.list}
\alias{example.map.list}
\docType{data}
\title{
an example REmap dataset list
}
\description{
an example REmap dataset list as GRanges generated from buildREmap
}
\usage{data(example.map.list)}
\format{
  The format is:
Formal class 'GRanges' [package "GenomicRanges"]
}
\examples{
library(REDseq)
data(example.map.list)
## maybe str(example.map.list) ; plot(example.map.list) ...
}
\keyword{datasets}
