\name{example.map}
\alias{example.map}
\docType{data}
\title{
an example REmap dataset
}
\description{
an example REmap dataset as RangedData generated from buildREmap
}
\usage{data(example.map)}
\format{
  The format is:
Formal class 'RangedData' [package "IRanges"]
}
\examples{
library(REDseq)
data(example.map)
## maybe str(example.map) ; plot(example.map) ...
}
\keyword{datasets}
