\name{example.REDseq.list}
\alias{example.REDseq.list}
\docType{data}
\title{
an example sequencing dataset list from a restoration enzyme digestion (RED) experiment
}
\description{
an example RED sequencing dataset list as a GRanges
}
\usage{data(example.REDseq.list)}
\format{
  The format is:
Formal class 'GRanges' [package "GenomicRanges"]
  }
\examples{
library(REDseq)
data(example.REDseq.list)
## maybe str(example.REDseq.list) ; plot(example.REDseq.list) ...
}
\keyword{datasets}
