\name{example.REDseq}
\alias{example.REDseq}
\docType{data}
\title{
an example sequencing dataset from a restoration enzyme digestion (RED) experiment
}
\description{
an example RED sequencing dataset as a GRanges
}
\usage{data(example.REDseq)}
\format{
  The format is:
Formal class 'GRanges' [package "GenomicRanges"]
  }
\examples{
library(REDseq)
data(example.REDseq)
## maybe str(example.REDseq) ; plot(example.REDseq) ...
}
\keyword{datasets}
